requirements.txt
### dependency
tqdm
dvc
pandas
numpy
PyYAML
joblib

### local packages -
# -e . # << uncomment this to install from setup.py
